refactored copyfiles into function to handle folders

+++ copyToDisk.lua
@@ -1,24 +1,19 @@
-local files = fs.find("*.lua")
+local args ={...}
+local path = args[0] or "/"
+copyFiles(path)
 
-for i = 1, #files, 1 do
-    print(files[i])
-    local newName = "disk/"..files[i]
-    if(fs.exists(newName)) then
-        fs.delete(newName)
+function copyFiles(path)
+    local files = fs.find(path.."*")
+    for i = 1, #files, 1 do
+        print(files[i])
+        if(fs.isDir(files[i])) then
+            copyFiles(path..files[i])
+        else
+            local newName = "disk"..path.."/"..files[i]
+            if(fs.exists(newName)) then
+                fs.delete(newName)
+            end
+            fs.copy(files[i], newName)
+        end
     end
-
-    fs.copy(files[i], newName)
-
-end
-
-local txtFiles = fs.find("*.txt")
-for i = 1, #txtFiles, 1 do
-    print(txtFiles[i])
-    local newName = "disk/"..txtFiles[i]
-    if(fs.exists(newName)) then
-        fs.delete(newName)
-    end
-
-    fs.copy(txtFiles[i], newName)
-
 end
